Move 'show templates' interaction into the dropdown menu
